+++ cutplace/rowio.py
@@ -16,6 +16,11 @@ format but not any :py:mod:`cutplace.fields` or :py:mod:`cutplace.checks`.
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
+
 import csv
 import datetime
 import io
@@ -296,6 +301,17 @@ def ods_rows(source_ods_path, sheet=1):
                 cell_value = ''
             else:
                 cell_value = text_p.text
+                if six.PY2:
+                    # HACK: It seems that under Python 2 ElementTree.find() returns a unicode string only of the value
+                    # actually contains non ASCII characters, and otherwise a binary string. To work around this we
+                    # check the result for binary strings and possibly convert them to uncicode strings assuming UTF-8
+                    # to be the internal encoding for the XML file. Ideally we would parse the XML header for the
+                    # encoding. Considering that Python 2 is on the way out, this just doesn't seem to be worth the
+                    # trouble right now.
+                    if isinstance(cell_value, six.binary_type):
+                        cell_value = six.text_type(cell_value, 'utf-8')
+                    else:
+                        assert isinstance(cell_value, six.text_type), 'cell_value=%r' % cell_value
             row.extend([cell_value] * repeated_count)
             location.advance_cell(repeated_count)
         yield row